Added consumers for rect/line. While they are basically the same, in terms of parameters, hopefully it's clear which is which.
Expanded search for BST to include a predicate on left/right, so you
can decide on which direction to take.

Also added radius onto the quadtree search parameter. The radius is
useful until there is a rewrite to determine if the quad actually
contains the x coordinate for a down search.